Moved API loading into Stack Navigation
