+++ screens/LoadConventionsScreen.js
@@ -1,9 +1,44 @@
 import React, { Component } from 'react';
-import ConventionListLoader from '../src/components/ConventionListLoader';
+import {Text} from 'react-native';
+import {withNavigationFocus} from 'react-navigation';
 
-export default class LoadConventionsScreen extends React.Component {
+import { goHome } from '../src/utils/Navigation';
 
-    render() {
-        return <ConventionListLoader onConventionsLoaded={this.props.onConventionsLoaded} />;
+import LoadConventionsContainer from '../src/components/LoadConventionsContainer';
+
+export class LoadConventionsScreen extends React.Component {
+  constructor(props){
+    super(props);
+
+    this.state = {
+      conventionList: null,
+      isLocalList: false,
+    };
+  }
+
+  handleConventionsLoaded = (conventions, isLocalData) => {
+    this.setState({
+      conventionList: conventions,
+      isLocalList: isLocalData,
+    });
+
+    this.props.navigation.navigate({
+      routeName: 'SelectConvention',
+      params: {conventionList: conventions, isLocalList: isLocalData},
+      key: 'SelectConvention'
+    });
+  }
+
+  render() {
+    //return <APILoader dataLoaded={this.props.dataLoaded} />;
+    if (!this.state.conventionList) {
+      return (
+        <LoadConventionsContainer onConventionsLoaded={this.handleConventionsLoaded} />
+      );
+    } else {
+      return <Text>Conventions loaded: {JSON.stringify(this.state.conventionList)}</Text>;
     }
+  }
 }
+
+export default withNavigationFocus(LoadConventionsScreen);
