+++ src/components/LoadConventionContainer.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
-import ConventionLoader from '../src/components/ConventionLoader';
+import ConventionLoader from './ConventionLoader';
 
-export default class LoadConventionScreen extends React.Component {
+export default class LoadConventionContainer extends React.Component {
 
     render() {
         return <ConventionLoader convention={this.props.conventionToLoad} onConventionLoaded={this.props.onConventionLoaded} onBack={this.props.onBack} />;
