+++ src/utils/Navigation.js
@@ -1,0 +1,3 @@
+export const goHome = (navigation) => {
+    navigation.navigate({ routeName: "Home", key: "Home" });
+}
